+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * ECS - Adjust restrict fund by location validation to accommodate central ordering. Refs UIOR-1235.
 * Use Save & close button label  stripes-component translation key. Refs UIOR-1240.
 * ECS - Add fields indicating the affiliation of locations and holdings for views. Refs UIOR-1276.
+* Custom Fields - Add filter components for purchase order lines. Refs UIOR-1237.
 
 ## [6.0.4](https://github.com/folio-org/ui-orders/tree/v6.0.4) (2024-04-25)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v6.0.3...v6.0.4)